requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
redis[asyncio]
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio
pytest-cov
anyio
ruff

